[Mission] delete missionStationLayer
